Allow multiple input-output pairs
Checking and generating a single input-output pair proved to be too
cumbersome on an actual code base
(https://github.com/admin-shell-io/aasx-package-explorer) since the
tool had to start up at each project individually.

This patch changes the command-line interface to accept multiple
input-output pairs so that the user is spared the start-up time
between the projects.