Add `--verbose` flag
A verbose output is good for, *e.g.*, tracking the progress and
verifying that all the files are checked as expected. For set projects
in which new files are rarely added verbose output actually makes
tracing errors very difficult. The user needs to go through a long list
of "OK" messages only to find one or two errors.

This patch introduces the command-line argument `--verbose` so that the
developer can choose which setting is preferred (progress bar *versus*
compact error messages).