Make difference reports more informative (#16)
This patch adds the diff to the report when there is a
difference between the actual and the generated test code. Usually, the
developer needs to just re-run the doctest generation to fix the check.

However, there might be very special circumstances causing the files to
be changed unepextedly. For example, if you reformat the code between
generating the test code and checking that it did not change.